VxMark: Handle primary elections in start voter session flow
In a general election, there's exactly one ballot style group for each
precinct or split. However, in a primary election, there are multiple -
one for each party. To handle this case, we add a second selection step
for the party when a poll worker is starting a new voter session.