VxAdmin: Label ballot styles by their districts/party in reports
In the VxMark ballot selection and VxAdmin manual tally contexts, ballot styles are
considered in conjunction with a precinct, so we can use precinct/split names to identify them. In the VxAdmin reporting context, a user may want to report on a single shared ballot style across precincts (e.g. if there are larger geographic areas that contain the precinct). So instead of identifying a ballot style/precinct combo, we use the list of districts associated with the ballot style as its label. Every ballot style has a unique list of districts, so this is the right unique label (along with the ballot style's party) when considering a ballot style across the entire election's geography.