don't build rust add-ons during the install phase
we're still building the rust add-on during the `checkout-and-install` step. previously, this was necessarily, because our `build` tooling didn't build those automatically. now it does, so they will be installed at build time. the builds aren't being cached anyway, I believe, so it's not saving any time. we could move to caching actual builds, but this current change is low hanging fruit.