Translate language names to every language
The eventual goal is to have the ballot language printed on the BMD
ballots in English (so that election workers can always tell what
language is on the ballot even if they don't recognize the text). In
order to do this, we need at the very least every ballot language to be
translated into English. I thought the most straightforward way to do
this would be to add a subkey for each language.