Human-friendly ballot style and language label on BMD ballots (#6377)
* Remove randomBallotId mocks

Leftover from https://github.com/votingworks/vxsuite/pull/6374

* Translate language names to every language

The eventual goal is to have the ballot language printed on the BMD
ballots in English (so that election workers can always tell what
language is on the ballot even if they don't recognize the text). In
order to do this, we need at the very least every ballot language to be
translated into English. I thought the most straightforward way to do
this would be to add a subkey for each language.

* Add "Language" title to app strings

For labelling the language on the ballot

* Human-friendly ballot style and language label on BMD ballots

- Change the ballot style label to match what's shown on screen (the
  precinct/split name and, for primaries, the party name)
- Add the ballot language, since that information was previously
  included in the ballot style ID

* Regenerate election fixtures

* Regenerate HMPB fixtures

* Update tests and snapshots

* Update BMD ballot test snapshots

* Update other snapshots due to election fixture hash change

* PR feedback