[VxDesign] Add Rust installation to deploy container image (#6412)
## Overview

The VxDesign deploy pipeline's been broken by a new transitive dependency on Rust tooling (which previously didn't need to be installed).

Adding a Rust installation layer to the Docker image to unblock. This also required removing `mark-scan` from the `dockerignore` list, since the Rust build fails otherwise, when it can't find all the projects referenced in the root `Cargo.toml`

## Testing Plan
- Via a Heroku review app