pdi-scanner: Disable auto-feeding after scan completes (#6386)
* pdi-scanner: Disable auto-feeding after scan completes.

We saw a few cases in the field where the logging indicated that a failed scan (likely due to the paper not being grabbed by the rollers) was followed quickly after by a second scan starting. In these cases, the state machine hadn't yet finished attempting to clear out the ballot from the first failed scan, and crashed when the second scan started, since it was unexpected.

To prevent scan events from occuring in rapid succession like this, we disable the auto-feeding immediately after receiving the scan end event from the PDI scanner. Unfortunately, we have to turn the feeder back on in order to eject ballots, but we do so immediately before sending the eject command, so it should be fairly safe.

* pdi-scanner: Clarify error condition when cropping scanned image

* Add expect assertion message