pdi-scanner: Retry connecting on Other error during connect
We saw a few machines that crashed due to an "Other" error during
connection. We haven't been able to reproduce, but it seems like it
might have been due to some sort of temporary blip. This PR turns that
error into a "disconnected" error, which will trigger reconnection logic
in the VxScan state machine. I also added a distinct error message so we
can see in the logs if this occurs in the future and evaluate whether
the reconnection logic helped.