Generate audit IDs and render one ballot for each ID
When a number of audit IDs is specified in the export task, instead of
generating a PDF for each combo of ballot type + ballot mode + ballot
style + precinct, just generate an official precinct ballot for a single
ballot style + precinct. Create a separate ballot PDF for each distinct
audit ID. This is a first pass to enable cert testing for this feature.