VxDesign: Export ballots with audit IDs (#6434)
* VxDesign: Add option to export ballots with audit IDs

When this option is selected, enable the user to specify how many audit IDs they want to generate. This will generate a separate ballot PDF for each audit ID. Also enable the flag in the system settings so that VxScan will know to look out for the audit IDs.

* Generate audit IDs and render one ballot for each ID

When a number of audit IDs is specified in the export task, instead of
generating a PDF for each combo of ballot type + ballot mode + ballot
style + precinct, just generate an official precinct ballot for a single
ballot style + precinct. Create a separate ballot PDF for each distinct
audit ID. This is a first pass to enable cert testing for this feature.

* Add audit ballot ID to default ballot template

- Encode the audit ID in the QR code metadata
- Print the audit ID with the other metadata in the footer (that way,
  you don't need to decode the QR code when auditing - you can just read
the ID from the footer)

* Rename getPdfFileName to getBallotPdfFileName

* Add test for rendering and interpreting ballot with audit ID

* VxDesign tests