fix(scan): stop using `ResizeStrategy::Fit`
We should always be getting images that are the correct size or approximately so. This is the first step toward stopping resizing of images, which is always wrong and possibly slow. Changes the logic for computing error to allow for a greater error in height than in width, since stretching can and does occur in the height dimension.