VxScan: Encrypt ballot audit IDs (#6455)
* libs/auth: Add AES-256 encryption functions

It was easiest to locate them here because the openssl utilities were
already built here, though I don't feel strongly about this choice.

* scan/backend: Generate and store secret key

* scan/backend: Encrypt ballot audit ID on export

The encryption logic is gated by the system settings flag.

Adding the ballotAuditId field to the sheet feeds it into the CVR
automatically, relying on previously built logic for VxCentralScan
imprinted IDs.

* Add script to decrypt the ballot audit IDs in a CVR export

It outputs a copy of each CVR file named with its ballot audit ID for
easy reference.

* Cryptography tests

* scan/backend test

* Switch to AES-256-EBC cipher

This simpler cipher is more easily securable if only limited to a single
block of data (128 bits), which should be plenty for our use case of
encrypting ballot serial numbers.

* PR feedback

* PR feedback