fix(scan): stop resizing scanned images (#6475)
* chore(fixtures): resize 300dpi to 200dpi

We scan at 200dpi, so there's no real sense having a 300dpi fixture around if we want to pass it to the scanning functions.

* fix(scan): stop using `ResizeStrategy::Fit`

We should always be getting images that are the correct size or approximately so. This is the first step toward stopping resizing of images, which is always wrong and possibly slow. Changes the logic for computing error to allow for a greater error in height than in width, since stretching can and does occur in the height dimension.

* chore(scan): remove `ResizeStrategy` and its uses

Since it became a singleton `struct` it has no effect. Inlines `ResizeStrategy::compute_error` at its one call site.

* chore(scan): remove resizing logic

* fix(scan): constrain width/height errors within ranges