fix(scan): recompute threshold after cropping (#6477)
We want it to be the case that we can always reinterpret an image and get the same result as the initial interpretation. This may not have been the case before because we used the pre-cropping threshold (which should be lower or equal) to determine the threshold we use. With just the cropped image it is impossible to recover that pre-cropping threshold, so we cannot reinterpret in exactly the same way. This change fixes that by re-computing the threshold after cropping.