fix(scan): prevent re-evaluating segments (#6517)
Sometimes, if the timing marks are mostly in a straight line, we might
end up with multiple segments to try that are actually the same. In the
best case this is doing repeated work, and in the worst case it's
causing us to miss a segment that would find all the timing marks
because we only evaluate so many pairs.