test: correctly pre-process bubble template image
Previously, we were not applying the same pre-processing to the bubble template image in test as we are in dev/prod. This resulted in very low bubble scores because very little of the grayscale template is actually pure black or white. Because the tests that care about bubble scores and votes are in the TypeScript layer, this was never a problem.