pdi-scanner: Ensure we always load image calibration tables (#6563)
Previously, we were only loading them in the case where we needed to
retry the initial commands after connect. This was happening so
consistently that I didn't notice the bug until just recently.

To mitigate this, I instead moved the request for the image calibration
tables to within the initial commands after connect.