refactor(scan): separate `contours` algorithm
I have been working to replace the `find_contours`-based timing mark detection algorithm, but it's a relatively large PR already and I thought it'd be good to break this part out separately.

The idea is that we move anything the new algorithm doesn't use into `timing_marks::contours`, leaving everything else the same as much as possible.