test: add proptests for parsing
The tests found some inputs that would crash the parse functions. For example:

```
Test failed: byte index 5 is not a char boundary; it is inside '¡' (bytes 4..6) of `0 ®¡AAAༀ𝒻0`.
minimal failing input: input = "0 ®¡AAAༀ𝒻0"
```

Because the offsets are bytes and not characters, you must be careful not to try to split a slice in the middle of a multi-byte UTF-8 character.