fix(scan): compute timing mark spacing properly
Rather than assuming that timing mark centers are always exactly separated by the sum of the width and height combined, calculate the actual separation we expect to find for each ballot size. This could be pretty important since the spacing actually varies because we currently use a fixed inset rather than fixing the spacing.