feat(scan): make `corners` timing mark algorithm the default (#6617)
* fix(scan): realign paper sizes

In 81eaa0c0be652aef334e951afb82b05d8f09c1e5 we replaced the 18" and 21" ballot sizes with a 19" ballot size, but we never updated the Rust side of the ballot interpreter to account for this. Oddly, the tests still passed when passing a 19" ballot for interpretation. We should figure out why that happened, but I don't think we need to block this on figuring that out.

* feat(scan): allow choosing the timing mark algorithm

* feat(scan): switch to the new timing mark algorithm by default