Replace Tone.js with HTMLAudioElement for AudioPlayer implementation
With our recent Electron version upgrade, we now have access to the
HTMLAudioElement's `preservesPitch` setting, so we can drastically
simplify the implementation. The diffs here won't be too useful, here,
since it's an apples-to-oranges change, so might be much easier/less
noisy to just view the code for the incoming changes on their own.