delete defunct logic to clear out search params
This logic seems to have no effect. For pollworkers we want to preserve
the search if a user clicks "cancel" from check-in. For election
managers we also want to preseve search if they click "close" on the voter
details page (not yet implemented). In any case, the current EM behavior
clears the search, so this was either never reached or redundant.