hmpb: Render root styles based on component supplied by template
Having the renderer be responsible for including `BaseStyles` was always a bit of a violation of the abstraction that templates specify content and styling while the renderer is just responsible for rendering it. It was a shortcut I put in place since all templates shared the same base styles. This started to break down when we needed to parameterize the styles based on ballot props (namely, adding `compact`, which changes font size and line height). In the future, it looks like we'll also need to change the fonts for upcoming NH state ballots. To get ahead of this, I added a new field to the ballot template to specify the base styles to use. The renderer will still only add this to the document head once at the beginning, but it will make it easier to customize these styles on a per-template basis moving forward. Plus, it patches up the separation of concerns shortcut.