feat(scan): change scale measurement to use full width (#6710)
This measurement should be generally quite close to the bottom
border-only measurement, but in practice will likely be a bit higher due
to it incorporating some of the skew that may be present toward the top
of the scan. Taking the median distance from the left to right timing
marks will mitigate that effect to some degree. On the TRR corpus the
difference in the measures is very close, about 0.08% average / 0.03%
median.