[pollbook] fix broken "Cancel" button after scan (#6705)
* invalidate barcode scan query after a successful scan

This prevents a race condition that happens when the user clicks
"cancel" or "close" to return to the search screen. The old barcode scan
value would still be cached in react-query, but the logic assumes a
scanned value can be read exactly once before it's consumed

* relax search when returning from pollworker voter check-in page

Improves UX because previously the invisible middle name, suffix, and
`exactMatch` params would be preserved in the search. This updates
search params after the scan and after the PW clicks "cancel" at the
voter check-in page. Only those search params that are visible and
editable by the user are kept in state.

* delete defunct logic to clear out search params

This logic seems to have no effect. For pollworkers we want to preserve
the search if a user clicks "cancel" from check-in. For election
managers we also want to preseve search if they click "close" on the voter
details page (not yet implemented). In any case, the current EM behavior
clears the search, so this was either never reached or redundant.