feat(fixtures): add API to make temp files/dirs (#6719)
`vxsuite` has lots of temporary files and directories created in tests, and is inconsistent about cleaning them up. When an effort is made to clean them up, there are often per-package ways of doing so that are effective to varying degrees. This is a step toward unifying tmpfile handling.

The idea is that all tests that need to create temporary files will use these functions to do so rather than using `tmp` (or `tmp-promise`). Why not use `tmp`? It refuses to delete temporary directories recursively on process exit when `setGracefulCleanup` is enabled, instead requiring that you call `dirSync` with `unsafeCleanup: true` and then manually invoke the `removeCallback` method on the returned `DirEntry`. This makes creating temporary files and directories in tests somewhat painful and encourages improper use such as just not cleaning them up or big hammers like `rm -rf /tmp/*`, which causes flaky tests and can potentially interfere with system services.

The next step after this is to integrate the `tmpdir` setup/teardown hooks from `@votingworks/fixtures` in any packages that create temporary files/dirs and replace any uses of `fileSync`/`dirSync`/`tmpNameSync` with the functions added here.