grout: Type inference for methods with no input that need context param
When a grout method has no input, that can be represented by no
parameter or a single param with type `undefined`. The latter is
necessary if the method needs to use the second param (API context)
supplied by middleware. E.g.: `getUser(input: undefined, context:
Context)`. When we call a method like this on the client, we want the
call site to look like this: `getUser()`, not `getUser(undefined)`.
That's what this type inference rule does.