design/backend: Refactor Auth0 client (#6731)
* Rename AuthClient to Auth0Client

Just to make it a bit clearer that it's responsible for interfacing with
Auth0 specifically.

* design/backend: Move `hasAccess` out of auth0 client

Our permissions rules don't actually have anything to do with Auth0. When we mock Auth0, it
shouldn't change our permissions rules.

* Make `userFromRequest` return a `User`

Hide the `Auth0User` type inside the client so the rest of VxDesign can
think in terms of the `User` type. This also allows us to remove the
`org()` method from the client, shrinking its surface area, which will
make it easier to mock.