VxDesign: Check authorization for all API endpoints (#6726)
* grout: Type inference for methods with no input that need context param

When a grout method has no input, that can be represented by no
parameter or a single param with type `undefined`. The latter is
necessary if the method needs to use the second param (API context)
supplied by middleware. E.g.: `getUser(input: undefined, context:
Context)`. When we call a method like this on the client, we want the
call site to look like this: `getUser()`, not `getUser(undefined)`.
That's what this type inference rule does.

* design/backend: Authorization checks for all API endpoints

- Adds middleware to load authenticated user info and store it in the
  API context
- Adds helper functions to check authorization to access a given org or
  election
- Adds middleware to check authorization for any API endpoints that have `orgId` or
  `electionId` input parameters
- For other API endpoints, calls helper functions in endpoint handler to check authorization

* Update ErrorBoundary to take message component instead of function

There was a lint rule violation with the previous approach of taking a
function prop. The lint rule claimed the function prop would cause a lot
of unnecessary re-renders and suggested using a component instead.

We'll use this prop to dynamically handle auth errors in the next
commit.

* design/frontend: Handle auth errors

- On unauthorized, redirect to login screen
- On forbidden, show a "page not found" message

* design/frontend: Update API calls and mocks based on signature changes

* design/backend: Tweak dev auth config

- Use the frontend port number for redirect URIs (this port number is
  already configured in Auth0)
- Proxy the frontend auth routes to the backend dev server

* Update existing backend tests

* Add backend tests for authorization

* Update frontend tests