grout: Change middleware functions into wrappers
Previously, API middleware could only execute before the RPC method
handler. Now, each middleware function is a wrapper, forming a nested
call chain with the RPC method at the center. This is a more powerful
pattern that allows us to create middleware that logs the result of the
call (including all the other middleware), or times the duration of an
API call.