dumb down our disk space monitoring
currently we have some extra elaborate disk space summary that tracks the maximum (of all time) available disk space in the store to use as the denominator in disk space figures. the purpose was to avoid the situation where you load up your device for the first time and it says "10GB / 110GB" used due to the code and other static files, much like a new iPhone would. the complexity is not worth it, though, and it turns out that apps other than VxAdmin are clearing the DB in full anyway, so it's not working on those apps.

this commit dumbs it down so we just ask the OS for the total and available disk space, even if the total includes disk space that will never be available to the application.