fix(scan): prevent losing image data in data race
We buffer the incoming packets from the scanner and, if a packet cannot
be handled by the code currently reading packets, put unhandled packets
into a list to be handled later. However, it's possible that the list
may include responses to requests that we're no longer able to handle
for whatever reason (e.g. a request timed out, but the response did
eventually come). Thus, when we make outgoing requests, because we only
expect there to be a single request/response communication happening at
a time with the scanner, we discard any incoming responses. We excluded
events from this because those are not in response to a request and
should not be discarded simply because we're going to make a request.
However, this code had a bug and was discarding image data.

The fix for this is relatively simple: make sure to keep image data when
discarding stale responses. However, in order to test this with the
scanner and verify that it was fixed manually, I needed to make other
changes to remove the blocking IO code that the "client" side of the
client/scanner pair was using. We'd switched over to using async with
tokio when switching from rusb to nusb, but the other half was still
using blocking IO. Mixing blocking and non-blocking IO must be done
carefully and is easy to get wrong, so it's simpler to just use
non-blocking IO (i.e. async) everywhere.