VxDesign: Decrypt CVR ballot audit IDs (#6869)
* design/backend: API method to decrypt CVR ballot audit IDs

Copies the meat of the libs/auth script with some input/output tweaks
(namely, taking in a ZIP file)

* design/frontend: Add form to export screen to upload CVR file for decryption

SLI and VX are currently the only users with access to this screen. We'll be moving SLI-specific
functionality to a bespoke screen shortly.

* ui: Allow uploading the same file twice in FileInputButton

Previously, the second upload would not trigger the onChange handler

* Backend test

* Frontend test

* Add unzip to Dockerfile