Proof-of-concept: Reduce memory usage when rendering ballots
The main memory bottleneck in the current code is that we keep a browser page open for each rendered ballot until we can extract the grid layouts, create the election definition, hash it, and inject that hash in the QR code.

I addressed this by doing two rendering passes: first rendering the ballot just to extract the grid layout and throwing it away. Then,
rendering the ballot again and injecting the QR code once we have the ballot hash ready.

I also noticed that even with this approach, there appeared to be some memory leakage from creating a new browser page for each ballot, even if I cleaned up the page after rendering. Since we render balots serially, it's safe to reuse the same browser page. Making that change seemed to stop the memory leak.