fix(scan): pass `Uint8Array` instead of `Buffer` (#6989)
The version of `pdfjs-dist` we updated to in #6756, v3.11.174, requires
the PDF data be passed as a `Uint8Array` rather than a `Buffer`. This is
annoying since TypeScript doesn't complain about a `Buffer` being passed
where a `Uint8Array` is expected since the former is a subclass of the
latter. #6756 updated most of the places where we were passing
`Buffer`s, but missed a couple that this commit addresses.