Properly error when programming cards configured for the incorrect environment (#7035)
* Read card details after programming

So that we properly error on cards configured for the incorrect
environment

* Simplify logic

* Add one more check for completeness though not strictly necessary

This check verifies that the card possesses the private key that
pairs with the card VotingWorks cert before proceeding with card
programming commands. This isn't strictly necessary from a security
perspective because we always perform this check during VxSuite
authentication but is nice from a completeness perspective.