VxDesign: Improve warning for error with syncing org cache
I realized that the previous guard to only sync the org cache on startup
in production was not working well in development. I originally thought
that running `pnpm db:reset-dev` would repopulate the org cache in
development, but that's only true if you run that command with
`AUTH_ENABLED=true` in its environment (it doesn't know how to check
.env.local).

So instead, I decided to change the sync-on-startup logic to run
whenever auth is enabled (in either prod or development). However, this
brings back the original issue: if you're using a prod db backup for
debugging but are pointed at the dev Auth0 instance, syncing will fail.

I added some logic to catch this error and log instructions for how to
easily work around this. Since debugging from a prod backup is an
uncommon use case, I'm ok with having to do a few steps to enable it.