VxDesign: Don't export sample ballots by default (#7261)
* VxDesign: Don't export sample ballots by default

Add a checkbox to include them. This reduces the number of ballots
exported in the default case by 1/3.

* Colocate payload schemas with task definitions

I forgot to update the schema, which caused a silent error only caught
by tests. With this change, it would be caught by the type system.

* Update backend tests

* Switch export audio control to a checkbox

The segmented button made it seem more like a persistent setting, which
it isn't. Also moved it next to the other export options.

* Remove unused type that was accidentally left around

* Update frontend tests