fix(scan): crop black around summary ballots
Closes #7213

When scanning on VxScan or VxCentralScan we may sometimes have extra
black areas around the ballot. This primarily happens at the header or
footer because the scanner turns on its image sensors before actually
catching the sheet (VxScan) or is expecting a longer ballot (i.e. for
long bubble ballots) on VxCentralScan.

For bubble ballots we crop these black areas around the page and so do
not have this problem. Summary ballots are still handled via a different
code path that did not have the cropping. This change ports the cropping
logic from the bubble ballot code to the summary ballot code, bringing
the implementations more in line.

Though we could likely reduce the search area for QR codes as a result
of this, we're opting not to touch that code out of an abundance of
caution and because we're going to replace it with the Rust code anyway.