VxDesign: Disallow precinct splits with the same district list
The point of precinct splits is to allow different combinations of
districts for different areas within a precinct. If two splits have the
same list of districts, there's no point in having those splits.

The ballot style generation code already assumes that this property is
true and relies on that assumption to create a valid set of ballot
styles.