VxDesign: Load and convert MS SEMS elections (#7400)
* Factor out helper to split candidate names

* Function to convert MS election files into a VX election

* Fixtures and tests for conversion

The fixtures are a selection of files from previous elections. I
replaced all of the town and candidate names with mock data.

* Extend loadElection API method to support MS SEMS election uploads

* Add feature flag to enable MS SEMS conversion

Currently only enabled for the VotingWorks org. I'll work out how to
flag MS orgs for this feature in a subsequent PR.

* Add modal for uploading MS SEMS or VX election files

Orgs without the feature flag enabled will still have the original UX
with no modal.

* PR feedback: Improve handling of ballot measures and add comments

* Catch database-level errors when inserting election

* Update existing design/backend tests

* design/backend integration tests

* Frontend tests

* fix typo