fix: ensure `Display` matches `FromStr` (#7572)
It appears clap converts the `default_value_t` value to a string using
`Display`, then parses it with `FromStr`. This means anything created by
`Display` for the default value has to parse with `FromStr`. This commit
updates both trait implementations to match and be simplified to allow
"enabled", "disabled", "enabled:10px", or "enabled:10".