fix(scan): stop accounting for possible timing mark cropping
We print the timing marks fairly inset now, so there shouldn't be any
instances of cropping due to scans not seeing to the edge of the paper.
This "correcting" behavior was causing issues when the timing marks
aren't cropped but are otherwise detected only partially, e.g. because
of a streak. In that situation the correction can make the effective
center of the timing mark even farther from its real value, leading to
inflated scores for empty bubbles. Sometimes these scores would exceed
the mark threshold, leading to spurious overvotes or unintended votes.

This change removes the "correction", leading to closer-to-zero scores
for empty bubbles when the timing marks are interrupted.