VxDesign/VxSuite: Fix SearchSelect menu so it's not hidden when used in a Modal (#7582)
* ui: Show SearchSelect in front of Modal when using a portal

SearchSelect supports using a portal to attach the menu to document.body
rather than the parent element, which is recommended in the docs for use
in modals. However, the react-modal z-index is 999, and the menu portal z-index
is only 10, so even when you use a portal, the menu shows up behind the
modal. By increasing the z-index of the portal to 1000, we fix this
problem.

* VxDesign: Use portal in org select modal

* Remove really tall modal workaround

* Remove overflow: visible workaround (which didn't actually work)

* Update tests