Update load election button
Always show the modal to simplify the code. I considered keeping the
special case for single-jurisdiction users with only a single file
format (e.g. NH users), but I figured it wasn't worthwhile since saving
clicks for this rare action isn't that important. Plus, NH users will
eventually have the option to upload elections in an NH-specific format
for state elections.

The other big change here is that I decided to use the jurisdiction's
stateCode directly rather than load feature flags. We'd need to add an
API method to load the feature flags for each jurisdiction (or add it to
the listJursidictions endpoint). I agonized over this a bit but since
there's only this one example to draw from, it felt to early to guess at
the right abstraction, so I decided to do the simplest thing to start.
Open to feedback and ideas on this though.