Remove jurisdiction ID deps from migrations
Since we won't have the special jurisdiction IDs in the env anymore, the
previous migrations that relied on them need to change. Since those
migrations have already run, we can actually remove all of parts that
specifically existed to migrate existing data in Auth0/the database and
assume we're starting from an empty database from now on (basically,
only to be used in development/test). I also moved the insertion of dev
data to a separate migration for clarity.