VxDesign: Base user features on organization, not jurisdictions (#7698)
* Base user feature flagging on organization, not jurisdiction

* Update tests

* Remove comment about removing env vars

Eventually, I'd like to pursue one of two paths:
- Move org-based feature flags into the database, so instead of
  specifying a few special orgs that get feature flags via env vars, we
can edit features for any org.
- Remove org-based feature flags entirely. Currently most of the feature
  flags are just to differentiate what VX support users can do from
everyone else, so we might not even need this once we have a first-class
notion of support users.

* Remove jurisdiction ID deps from migrations

Since we won't have the special jurisdiction IDs in the env anymore, the
previous migrations that relied on them need to change. Since those
migrations have already run, we can actually remove all of parts that
specifically existed to migrate existing data in Auth0/the database and
assume we're starting from an empty database from now on (basically,
only to be used in development/test). I also moved the insertion of dev
data to a separate migration for clarity.